Add solution for problem `A - C+=` Link: https://codeforces.com/contest/1368/problem/A Tags: brute force, greedy, implementation, math, *800 Ref: https://codeforces.com/contest/1368/submission/127644935

+++ codeforces/1368/A.cpp
@@ -28,27 +28,16 @@ const int MAX=100005;
 
 void solve()
 {
-ll a,b,n;
-cin>>a>>b>>n;
-ll ans=0;
-while(a<=n && b<=n)
-{
-  
-  if(a<b)
-    
-    a+=b;
-  
-  else 
-    b+=a;
-    ans++;
-}
-
-    cout<<ans<<"\n";
-  
-  
-
-  
-
+ll a, b, n;
+ cin >> a >> b >> n;
+		int ans = 0;
+		while (a <= n && b <= n) {
+			if (a < b) a += b;
+			else b += a;
+			ans++;
+		}
+		cout << ans << '\n';
+	
  
   }
